Moved autoreconnect to WelcomeMessage.tsx
