+++ src/components/pages/StudentsPage/WelcomeMessage.tsx
@@ -1,7 +1,9 @@
 /** @jsxImportSource @emotion/react */
 
+import { useEffect } from 'react';
 import { Box, Typography } from '@mui/material';
 import Image from 'next/image';
+import type { Socket } from 'socket.io-client';
 
 import Link from '@components/shared/Link';
 
@@ -9,15 +11,62 @@ interface WelcomeMessageProps {
   pin: string;
   removedFromClass: boolean;
   socketId: string;
+  socket: Socket;
   studentName: string;
   isMobile: boolean;
+  addStudentToGameroom: (
+    studentName: string,
+    pin: string,
+    updateStageToLobby?: boolean,
+  ) => void;
 }
 
 export default function WelcomeMessage({
+  pin,
   removedFromClass,
+  socket,
+  socketId,
   studentName,
   isMobile,
+  addStudentToGameroom,
 }: WelcomeMessageProps) {
+  useEffect(() => {
+    if (!socket || removedFromClass || !studentName || !pin) return;
+
+    async function reconnectToGameroom() {
+      const apiUrl = `${process.env.NEXT_PUBLIC_SERVER_URL}/api/v1`;
+
+      try {
+        const getResponse = await fetch(
+          `${apiUrl}/classrooms/${pin}/studentSockets/${socketId}`,
+          { method: 'GET' },
+        );
+        const { isStudentInsideClassroom } = await getResponse.json();
+
+        if (!isStudentInsideClassroom) {
+          const updateStageToLobby = false;
+          addStudentToGameroom(studentName, pin, updateStageToLobby);
+        }
+      } catch (e) {
+        console.log(e);
+      }
+    }
+
+    // Keep students in the same classroom after a brief reconnect while in lobby.
+    socket.on('connect', reconnectToGameroom);
+
+    return () => {
+      socket.off('connect', reconnectToGameroom);
+    };
+  }, [
+    addStudentToGameroom,
+    pin,
+    removedFromClass,
+    socket,
+    socketId,
+    studentName,
+  ]);
+
   return (
     <Box textAlign='center'>
       <Image
